refactor(kubernautagent): split cmd/kubernautagent/main.go into focused files
main.go shrank from 1,920 to 345 lines (main(), loadStartupConfig,
resolveLLMCredentials only). Extracted:
- bootstrap.go: LLM client + alignment stack + investigation runner wiring
- health.go: health/readiness/config handlers + K8s infra bootstrap
- datastorage.go: DataStorage client + enricher/sanitizer/audit-store wiring
- toolregistry.go: tool registry construction + fleet tool registration
- routes.go: API route tree + MCP interactive handler wiring

Pure structural move within the same package; zero behavior change, no
exported-API change. Verified with go build, go vet, gofmt, and the full
cmd/kubernautagent unit-test suite (Wave 4, GO-ANTIPATTERN-AUDIT-2026-07-01).
