+++ cmd/kubernautagent/main.go
@@ -18,82 +18,31 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
 	"net/http"
-	"net/http/pprof"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"strings"
 	"sync/atomic"
 	"syscall"
 	"time"
 
 	"github.com/go-chi/chi/v5"
-	chimiddleware "github.com/go-chi/chi/v5/middleware"
 	"github.com/go-logr/logr"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"go.uber.org/zap"
-	k8sruntime "k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/client-go/discovery/cached/memory"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/kubernetes"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/restmapper"
-	ctrl "sigs.k8s.io/controller-runtime"
 
-	remediationv1 "github.com/jordigilh/kubernaut/api/remediation/v1alpha1"
-	"github.com/jordigilh/kubernaut/pkg/agentclient"
-	sharedaudit "github.com/jordigilh/kubernaut/pkg/audit"
-	ogenclient "github.com/jordigilh/kubernaut/pkg/datastorage/ogen-client"
 	"github.com/jordigilh/kubernaut/pkg/kubernautagent/llm"
-	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 	auth "github.com/jordigilh/kubernaut/pkg/shared/auth"
-	"github.com/jordigilh/kubernaut/pkg/shared/hotreload"
-	sharedtls "github.com/jordigilh/kubernaut/pkg/shared/tls"
-	sharedtransport "github.com/jordigilh/kubernaut/pkg/shared/transport"
-	"github.com/jordigilh/kubernaut/pkg/shared/types"
 
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/alignment"
-	alignprompt "github.com/jordigilh/kubernaut/internal/kubernautagent/alignment/prompt"
-	kaapi "github.com/jordigilh/kubernaut/internal/kubernautagent/api"
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/audit"
 	kaconfig "github.com/jordigilh/kubernaut/internal/kubernautagent/config"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/credentials"
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/enrichment"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/investigator"
-	mcpkg "github.com/jordigilh/kubernaut/internal/kubernautagent/mcp"
-	mcpadapters "github.com/jordigilh/kubernaut/internal/kubernautagent/mcp/adapters"
-	mcptools "github.com/jordigilh/kubernaut/internal/kubernautagent/mcp/tools"
-	kametrics "github.com/jordigilh/kubernaut/internal/kubernautagent/metrics"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/parser"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/prompt"
 	karbac "github.com/jordigilh/kubernaut/internal/kubernautagent/rbac"
 	kaserver "github.com/jordigilh/kubernaut/internal/kubernautagent/server"
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/session"
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/tools/custom"
-	"github.com/jordigilh/kubernaut/internal/version"
-	dsschema "github.com/jordigilh/kubernaut/pkg/datastorage/schema"
-	fleetclient "github.com/jordigilh/kubernaut/pkg/fleet/mcpclient"
-	fleetregistry "github.com/jordigilh/kubernaut/pkg/fleet/registry"
-	amtools "github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/alertmanager"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/investigation"
-	k8stools "github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/k8s"
-	logtools "github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/logs"
-	promtools "github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/prometheus"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/registry"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/sanitization"
-	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/summarizer"
 	kubelog "github.com/jordigilh/kubernaut/pkg/log"
-	wfclient "github.com/jordigilh/kubernaut/pkg/workflowexecution/client"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/client-go/rest"
-	metricsclient "k8s.io/metrics/pkg/client/clientset/versioned"
-	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func main() {
@@ -406,1543 +355,4 @@ func resolveLLMCredentials(cfg *kaconfig.Config, llmRuntime *kaconfig.LLMRuntime
 		logger.Info("OAuth2 credentials resolved from mounted Secret",
 			"credentialsDir", cfg.AI.LLM.OAuth2.CredentialsDir)
 	}
-}
-
-// buildLLMClients constructs the primary SwappableClient plus any per-phase
-// SwappableClient overrides configured in llmRuntime.PhaseModels. Terminates
-// the process (os.Exit(1)) on unrecoverable client construction failures,
-// matching the original inline main() behavior
-// (GO-ANTIPATTERN-AUDIT-2026-07-01 Phase 4f).
-func buildLLMClients(cfg *kaconfig.Config, llmRuntime *kaconfig.LLMRuntimeConfig, logger logr.Logger) (*llm.SwappableClient, map[katypes.Phase]*llm.SwappableClient) {
-	llmClient, err := buildLLMClientFromConfig(context.Background(), mergeLLMConfig(cfg.AI.LLM, llmRuntime))
-	if err != nil {
-		logger.Error(err, "failed to create LLM client", "provider", cfg.AI.LLM.Provider)
-		os.Exit(1)
-	}
-
-	swappable, err := llm.NewSwappableClient(llmClient, llmRuntime.Model, llm.RuntimeParams{
-		Temperature:    llmRuntime.Temperature,
-		TimeoutSeconds: llmRuntime.TimeoutSeconds,
-		MaxRetries:     llmRuntime.MaxRetries,
-	})
-	if err != nil {
-		logger.Error(err, "failed to create swappable LLM client")
-		os.Exit(1)
-	}
-
-	// #1470: Build per-phase SwappableClients. The resolver is created after
-	// the alignment evaluator, so that the PinDecorator can be set at
-	// construction time.
-	phaseSwappables := make(map[katypes.Phase]*llm.SwappableClient)
-	for phaseName, override := range llmRuntime.PhaseModels {
-		phaseLLM, phaseRT := llmRuntime.EffectivePhaseConfig(phaseName, cfg.AI.LLM, *llmRuntime)
-		phaseClient, phaseErr := buildLLMClientFromConfig(context.Background(), mergeLLMConfig(phaseLLM, &phaseRT))
-		if phaseErr != nil {
-			logger.Error(phaseErr, "failed to build phase LLM client",
-				"phase", phaseName, "model", override.Model)
-			os.Exit(1)
-		}
-		phaseSw, phaseSwErr := llm.NewSwappableClient(phaseClient, phaseRT.Model, llm.RuntimeParams{
-			Temperature:    phaseRT.Temperature,
-			TimeoutSeconds: phaseRT.TimeoutSeconds,
-			MaxRetries:     phaseRT.MaxRetries,
-		})
-		if phaseSwErr != nil {
-			logger.Error(phaseSwErr, "failed to create phase SwappableClient",
-				"phase", phaseName)
-			os.Exit(1)
-		}
-		phaseSwappables[katypes.Phase(phaseName)] = phaseSw
-		logger.Info("per-phase LLM client initialized",
-			"phase", phaseName, "model", phaseRT.Model, "override_model", override.Model)
-	}
-
-	return swappable, phaseSwappables
-}
-
-// buildAlignmentStack resolves the shadow-agent alignment-check
-// configuration and, when enabled, constructs the dedicated (or shared)
-// shadow LLM client and wraps the primary LLM/tool registry in
-// alignment-aware proxies. Terminates the process on a fail-closed shadow
-// client construction failure, matching the original inline main() behavior
-// (GO-ANTIPATTERN-AUDIT-2026-07-01 Phase 4f).
-func buildAlignmentStack(
-	cfg *kaconfig.Config,
-	llmRuntime *kaconfig.LLMRuntimeConfig,
-	instrumentedLLM llm.Client,
-	reg *registry.Registry,
-	auditStore audit.AuditStore,
-	logger logr.Logger,
-) (effectiveLLM llm.Client, effectiveReg registry.ToolRegistry, alignEvaluator *alignment.Evaluator, alignCfg kaconfig.AlignmentCheckConfig) {
-	effectiveLLM = instrumentedLLM
-	effectiveReg = reg
-	alignCfg = resolveAlignmentCheckConfig(cfg)
-	if !alignCfg.Enabled {
-		return effectiveLLM, effectiveReg, nil, alignCfg
-	}
-
-	var shadowClient llm.Client
-	if alignCfg.LLM == nil {
-		shadowClient = instrumentedLLM
-		logger.Error(nil, "shadow agent shares investigation LLM client — shadow requests will compete with primary investigation; configure ai.alignmentCheck.llm for dedicated shadow model")
-	} else {
-		alignStaticCfg, alignRtCfg := alignCfg.EffectiveLLM(cfg.AI.LLM, *llmRuntime)
-		raw, alignErr := buildLLMClientFromConfig(context.Background(), mergeLLMConfig(alignStaticCfg, &alignRtCfg))
-		if alignErr != nil {
-			logger.Error(alignErr, "alignment check LLM client failed (fail-closed): alignment is enabled but shadow client unavailable")
-			os.Exit(1)
-		} else {
-			shadowClient = llm.NewInstrumentedClient(raw)
-			logger.Info("shadow agent using dedicated LLM client", "model", alignRtCfg.Model)
-		}
-	}
-	if shadowClient != nil {
-		alignEvaluator = alignment.NewEvaluator(shadowClient, alignment.EvaluatorConfig{
-			Timeout:               alignCfg.Timeout,
-			MaxStepTokens:         alignCfg.MaxStepTokens,
-			MaxRetries:            alignCfg.MaxRetries,
-			MaxConversationTokens: alignCfg.GroundingReview.MaxConversationTokens,
-		}, alignprompt.SystemPrompt(), alignment.WithLogger(logger), alignment.WithAuditStore(auditStore))
-		effectiveLLM = alignment.NewLLMProxy(instrumentedLLM)
-		effectiveReg = alignment.NewToolProxy(reg)
-		logger.Info("shadow agent alignment check enabled (shadow LLM audit active: request/response events will be emitted per step)")
-	}
-	return effectiveLLM, effectiveReg, alignEvaluator, alignCfg
-}
-
-// investigationRunnerParams groups the dependencies needed to build the
-// investigation stack (investigator, session store/manager, ogen server).
-// Extracted per AGENTS.md's 8+-param Options-pattern rule.
-type investigationRunnerParams struct {
-	cfg             *kaconfig.Config
-	llmRuntime      *kaconfig.LLMRuntimeConfig
-	swappable       *llm.SwappableClient
-	phaseSwappables map[katypes.Phase]*llm.SwappableClient
-	promptBuilder   *prompt.Builder
-	resultParser    *parser.ResultParser
-	phaseTools      katypes.PhaseToolMap
-	enricher        *enrichment.Enricher
-	auditStore      audit.AuditStore
-	effectiveLLM    llm.Client
-	effectiveReg    registry.ToolRegistry
-	alignEvaluator  *alignment.Evaluator
-	alignCfg        kaconfig.AlignmentCheckConfig
-	infra           *k8sInfra
-	sanitizer       *sanitization.Pipeline
-	anomalyDetector *investigator.AnomalyDetector
-	catalogFetcher  investigator.CatalogFetcher
-	summarizer      *summarizer.Summarizer
-	logger          logr.Logger
-}
-
-// investigationStack groups the constructed investigation-stack components
-// that main() needs after buildInvestigationRunner: the investigator itself
-// (needed by buildMCPHandler), the metrics/audit/session infrastructure, and
-// the ogen server mounted on the router.
-type investigationStack struct {
-	agentMetrics      *kametrics.Metrics
-	instrumentedAudit audit.AuditStore
-	phaseResolver     *investigator.DefaultPhaseResolver
-	inv               *investigator.Investigator
-	store             *session.Store
-	mgr               *session.Manager
-	ogenSrv           *agentclient.Server
-}
-
-// buildInvestigationRunner wires the investigator, the alignment wrapper
-// (when enabled), the session store/manager, and the ogen server. Terminates
-// the process on unrecoverable wiring failures, matching the original inline
-// main() behavior (GO-ANTIPATTERN-AUDIT-2026-07-01 Phase 4f).
-func buildInvestigationRunner(p investigationRunnerParams) *investigationStack {
-	agentMetrics := kametrics.NewMetrics()
-
-	instrumentedAudit := audit.NewInstrumentedAuditStore(p.auditStore, agentMetrics.RecordAuditEventEmitted)
-
-	var scopeResolver investigator.ScopeResolver
-	if p.infra != nil {
-		scopeResolver = investigator.NewMapperScopeResolver(p.infra.mapper)
-	}
-
-	// #1470: Build the PhaseResolver with the PinDecorator (if alignment is enabled).
-	var pinDecorator func(llm.Client) llm.Client
-	if p.alignEvaluator != nil {
-		pinDecorator = func(pinned llm.Client) llm.Client {
-			return alignment.NewLLMProxy(llm.NewInstrumentedClient(pinned))
-		}
-	}
-	phaseResolver := investigator.NewDefaultPhaseResolver(p.swappable, pinDecorator)
-	for phase, phaseSw := range p.phaseSwappables {
-		phaseResolver.SetPhaseSwappable(phase, phaseSw)
-	}
-
-	invCfg := investigator.Config{
-		Client:        p.effectiveLLM,
-		Builder:       p.promptBuilder,
-		ResultParser:  p.resultParser,
-		Enricher:      p.enricher,
-		AuditStore:    instrumentedAudit,
-		Logger:        p.logger,
-		MaxTurns:      p.cfg.AI.Investigation.MaxTurns,
-		PhaseTools:    p.phaseTools,
-		Registry:      p.effectiveReg,
-		ModelName:     p.llmRuntime.Model,
-		Swappable:     p.swappable,
-		ScopeResolver: scopeResolver,
-		Metrics:       agentMetrics,
-		PhaseResolver: phaseResolver,
-		PinDecorator:  pinDecorator,
-		Pipeline: investigator.Pipeline{
-			Sanitizer:         p.sanitizer,
-			AnomalyDetector:   p.anomalyDetector,
-			CatalogFetcher:    p.catalogFetcher,
-			Summarizer:        p.summarizer,
-			MaxToolOutputSize: p.cfg.AI.Summarizer.MaxToolOutputSize,
-		},
-	}
-	inv := investigator.New(invCfg)
-
-	var investigationRunner kaserver.InvestigationRunner = inv
-	if p.alignEvaluator != nil {
-		wrapper, wrapErr := alignment.NewInvestigatorWrapper(alignment.InvestigatorWrapperConfig{
-			Inner:                 inv,
-			Evaluator:             p.alignEvaluator,
-			VerdictTimeout:        p.alignCfg.VerdictTimeout,
-			AuditStore:            instrumentedAudit,
-			Logger:                p.logger,
-			Mode:                  p.alignCfg.Mode,
-			CanaryForceEscalation: p.alignCfg.Canary.ForceEscalation,
-			GroundingEnabled:      p.alignCfg.GroundingReview.Enabled,
-		})
-		if wrapErr != nil {
-			p.logger.Error(wrapErr, "failed to create alignment wrapper")
-			os.Exit(1)
-		}
-		investigationRunner = wrapper
-	}
-
-	store := session.NewStore(p.cfg.Runtime.Session.TTL,
-		session.WithLogger(p.logger.WithName("session-store")),
-		session.WithMaxConcurrent(p.cfg.Runtime.Session.MaxConcurrentInvestigations),
-	)
-	mgr := session.NewManager(store, p.logger, instrumentedAudit, agentMetrics)
-
-	handler := kaserver.NewHandler(mgr, investigationRunner, p.logger, agentMetrics)
-
-	ogenSrv, err := agentclient.NewServer(handler)
-	if err != nil {
-		p.logger.Error(err, "failed to create ogen server")
-		os.Exit(1)
-	}
-
-	return &investigationStack{
-		agentMetrics:      agentMetrics,
-		instrumentedAudit: instrumentedAudit,
-		phaseResolver:     phaseResolver,
-		inv:               inv,
-		store:             store,
-		mgr:               mgr,
-		ogenSrv:           ogenSrv,
-	}
-}
-
-// wireHotReload configures conditional server TLS (with hot-reload), the
-// LLM-runtime config watcher, the OCP TLS security profile, and the
-// CA-file watcher. Mutates httpServer.TLSConfig in place when TLS is
-// enabled. Terminates the process on unrecoverable wiring failures, matching
-// the original inline main() behavior.
-//
-// Returns a single cleanup function that stops every watcher that was
-// successfully started; the caller must defer it in its own scope so the
-// watchers live for the server's lifetime, not just this function's call
-// (GO-ANTIPATTERN-AUDIT-2026-07-01 Phase 4f — same pattern as
-// watchLoopState.stopEAWatcher in Phase 4e).
-func wireHotReload(
-	ctx context.Context,
-	cfg *kaconfig.Config,
-	httpServer *http.Server,
-	llmRuntimePath string,
-	swappable *llm.SwappableClient,
-	phaseResolver *investigator.DefaultPhaseResolver,
-	logger logr.Logger,
-) func() {
-	var stoppers []func()
-
-	// Issue #493: Conditional TLS for the HTTP server
-	// Issue #756: CertReloader enables hot-reload of server certificates
-	var certReloader *sharedtls.CertReloader
-	if cfg.Runtime.Server.TLS.Enabled() {
-		isTLS, reloader, tlsErr := sharedtls.ConfigureConditionalTLS(httpServer, cfg.Runtime.Server.TLS.CertDir)
-		if tlsErr != nil {
-			logger.Error(tlsErr, "Failed to configure TLS")
-			os.Exit(1)
-		}
-		if isTLS {
-			certReloader = reloader
-			logger.Info("TLS configured for HTTP server", "certDir", cfg.Runtime.Server.TLS.CertDir)
-		}
-	}
-
-	// Issue #756: Wire FileWatcher for server cert hot-reload
-	if certReloader != nil {
-		certWatcher, watchErr := hotreload.NewFileWatcher(
-			filepath.Join(cfg.Runtime.Server.TLS.CertDir, "tls.crt"),
-			certReloader.ReloadCallback,
-			logger.WithName("cert-reloader"),
-		)
-		if watchErr != nil {
-			logger.Error(watchErr, "Failed to create cert file watcher")
-			os.Exit(1)
-		}
-		if err := certWatcher.Start(ctx); err != nil {
-			logger.Error(err, "Failed to start cert file watcher")
-			os.Exit(1)
-		}
-		stoppers = append(stoppers, certWatcher.Stop)
-	}
-
-	// Issue #916: Wire FileWatcher for LLM runtime config hot-reload
-	rtCallback := llmRuntimeReloadCallback(cfg, swappable, logger, phaseResolver)
-	rtWatcher, rtWatchErr := hotreload.NewFileWatcher(
-		llmRuntimePath,
-		rtCallback,
-		logger.WithName("llm-runtime-reloader"),
-	)
-	if rtWatchErr != nil {
-		logger.Info("llm runtime file watcher not started", "error", rtWatchErr)
-	} else {
-		if err := rtWatcher.Start(ctx); err != nil {
-			logger.Info("llm runtime file watcher failed to start", "error", err)
-		} else {
-			stoppers = append(stoppers, rtWatcher.Stop)
-			logger.Info("llm runtime hot-reload enabled (#916)", "path", llmRuntimePath)
-		}
-	}
-
-	// Issue #748: Load OCP TLS security profile from config before any TLS setup
-	if err := sharedtls.SetDefaultSecurityProfileFromConfig(cfg.Runtime.Server.TLSProfile); err != nil {
-		logger.Error(err, "Invalid TLS security profile in config, using default TLS 1.2")
-	} else if cfg.Runtime.Server.TLSProfile != "" {
-		logger.Info("TLS security profile active", "profile", cfg.Runtime.Server.TLSProfile)
-	}
-
-	// Issue #756: Start CA file watcher for client-side TLS hot-reload
-	caWatcher, caWatchErr := sharedtls.StartCAFileWatcher(ctx, logger)
-	if caWatchErr != nil {
-		logger.Error(caWatchErr, "Failed to start CA file watcher")
-		os.Exit(1)
-	}
-	if caWatcher != nil {
-		stoppers = append(stoppers, caWatcher.Stop)
-	}
-
-	return func() {
-		for _, stop := range stoppers {
-			stop()
-		}
-	}
-}
-
-func shutdownTimeout(cfg *kaconfig.Config) time.Duration {
-	if cfg.Runtime.Shutdown.DrainSeconds > 0 {
-		return time.Duration(cfg.Runtime.Shutdown.DrainSeconds) * time.Second
-	}
-	return 30 * time.Second
-}
-
-func shutdownServer(ctx context.Context, srv *http.Server, name string, logger logr.Logger) {
-	if err := srv.Shutdown(ctx); err != nil {
-		logger.Error(err, "server shutdown error", "server", name)
-	}
-}
-
-func healthHandler(w http.ResponseWriter, _ *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	_ = json.NewEncoder(w).Encode(map[string]string{
-		"status":    "healthy",
-		"timestamp": time.Now().UTC().Format(time.RFC3339),
-	})
-}
-
-// readinessHandler returns an http.HandlerFunc that performs real
-// dependency checks instead of returning a static 200. Checks:
-//   - shutdownFlag: set after SIGTERM/SIGINT received; makes probe fail
-//     so k8s stops routing traffic during graceful shutdown.
-//   - apiServerReady: set once the main API server goroutine is about to
-//     start listening. Guards against the readiness probe reporting ready
-//     before the main API server (auth middleware + JWKS pre-warm) has
-//     finished initializing, which would otherwise let traffic hit a port
-//     that isn't accepting connections yet.
-//   - swappable: verifies the LLM client has a non-empty model name
-//     (proxy for "LLM client was successfully initialized").
-//   - ds: if non-nil, verifies the ogen client is initialized (DS is
-//     a soft dependency — nil ds means DS is intentionally unconfigured).
-//   - interactive: reports the interactive mode status (#891). This is
-//     informational (does not fail the probe) since autonomous mode
-//     continues to function even when interactive is soft-disabled.
-func readinessHandler(shutdownFlag, apiServerReady *int32, swappable *llm.SwappableClient, ds *dsClients, interactive *karbac.InteractiveReadiness) http.HandlerFunc {
-	return func(w http.ResponseWriter, _ *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-
-		if atomic.LoadInt32(shutdownFlag) != 0 {
-			w.WriteHeader(http.StatusServiceUnavailable)
-			_ = json.NewEncoder(w).Encode(map[string]string{
-				"status": "not_ready",
-				"reason": "shutting_down",
-			})
-			return
-		}
-
-		if atomic.LoadInt32(apiServerReady) == 0 {
-			w.WriteHeader(http.StatusServiceUnavailable)
-			_ = json.NewEncoder(w).Encode(map[string]string{
-				"status": "not_ready",
-				"reason": "api_server_starting",
-			})
-			return
-		}
-
-		if swappable.ModelName() == "" {
-			w.WriteHeader(http.StatusServiceUnavailable)
-			_ = json.NewEncoder(w).Encode(map[string]string{
-				"status": "not_ready",
-				"reason": "llm_client_not_configured",
-			})
-			return
-		}
-
-		if ds != nil && ds.ogenClient == nil {
-			w.WriteHeader(http.StatusServiceUnavailable)
-			_ = json.NewEncoder(w).Encode(map[string]string{
-				"status": "not_ready",
-				"reason": "datastorage_client_unavailable",
-			})
-			return
-		}
-
-		resp := map[string]string{
-			"status":           "ready",
-			"interactive_mode": interactive.StatusString(),
-		}
-		if reason := interactive.Reason(); reason != "" {
-			resp["interactive_reason"] = reason
-		}
-		_ = json.NewEncoder(w).Encode(resp)
-	}
-}
-
-func configHandler(cfg *kaconfig.Config, swappable *llm.SwappableClient) http.HandlerFunc {
-	return func(w http.ResponseWriter, _ *http.Request) {
-		model := swappable.ModelName()
-		sanitized := map[string]interface{}{
-			"service":     "kubernaut-agent",
-			"version":     version.Version,
-			"llm_model":   model,
-			"session_ttl": cfg.Runtime.Session.TTL.String(),
-		}
-		w.Header().Set("Content-Type", "application/json")
-		_ = json.NewEncoder(w).Encode(sanitized)
-	}
-}
-
-// startHealthAndMetricsServers builds and starts (in background goroutines)
-// the health/readiness/config/admin HTTP server and the Prometheus metrics
-// server. These are started before API route setup so liveness/readiness
-// probes are served even while the JWKS pre-warm (up to 15s) blocks inside
-// newAuthMiddleware — otherwise the liveness probe kills the pod before the
-// health server ever starts.
-func startHealthAndMetricsServers(cfg *kaconfig.Config, atomicLevel zap.AtomicLevel, swappable *llm.SwappableClient, ds *dsClients, interactiveReadiness *karbac.InteractiveReadiness, shutdownFlag, apiServerReady *int32, logger logr.Logger) (*http.Server, *http.Server) {
-	healthMux := http.NewServeMux()
-	healthMux.HandleFunc("/healthz", healthHandler)
-	healthMux.HandleFunc("/readyz", readinessHandler(shutdownFlag, apiServerReady, swappable, ds, interactiveReadiness))
-	healthMux.HandleFunc("/config", configHandler(cfg, swappable))
-	if !cfg.Runtime.Server.DisableAdminEndpoints {
-		healthMux.Handle("/admin/loglevel", atomicLevel)
-	}
-	healthMux.HandleFunc("/openapi.json", kaapi.SpecHandler())
-	if !cfg.Runtime.Server.DisableProfiling {
-		healthMux.HandleFunc("/debug/pprof/", pprof.Index)
-		healthMux.HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
-		healthMux.HandleFunc("/debug/pprof/profile", pprof.Profile)
-		healthMux.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
-		healthMux.HandleFunc("/debug/pprof/trace", pprof.Trace)
-	}
-	healthServer := &http.Server{
-		Addr:              cfg.Runtime.Server.HealthAddr,
-		Handler:           healthMux,
-		ReadHeaderTimeout: 5 * time.Second,
-		ReadTimeout:       10 * time.Second,
-		WriteTimeout:      10 * time.Second,
-	}
-	metricsMux := http.NewServeMux()
-	metricsMux.Handle("/metrics", promhttp.Handler())
-	metricsServer := &http.Server{
-		Addr:              cfg.Runtime.Server.MetricsAddr,
-		Handler:           metricsMux,
-		ReadHeaderTimeout: 5 * time.Second,
-		ReadTimeout:       5 * time.Second,
-		WriteTimeout:      10 * time.Second,
-		IdleTimeout:       30 * time.Second,
-	}
-
-	go func() {
-		logger.Info("health server listening", "addr", cfg.Runtime.Server.HealthAddr)
-		if err := healthServer.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			logger.Error(err, "health server error")
-		}
-	}()
-	go func() {
-		logger.Info("metrics server listening", "addr", cfg.Runtime.Server.MetricsAddr)
-		if err := metricsServer.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			logger.Error(err, "metrics server error")
-		}
-	}()
-
-	return healthServer, metricsServer
-}
-
-// detectNamespace reads the pod's namespace from the mounted ServiceAccount.
-func detectNamespace() string {
-	data, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
-	if err == nil && len(data) > 0 {
-		return strings.TrimSpace(string(data))
-	}
-	return "kubernaut-system"
-}
-
-// k8sInfra holds shared Kubernetes clients created once and reused by
-// the tool registry, enricher, and custom tools.
-type k8sInfra struct {
-	kubeConfig *rest.Config
-	clientset  *kubernetes.Clientset
-	dynClient  dynamic.Interface
-	mapper     meta.RESTMapper
-}
-
-// initK8sInfra creates the shared Kubernetes clients. Returns nil when
-// running outside a cluster (e.g. local development).
-func initK8sInfra(logger logr.Logger) *k8sInfra {
-	kubeConfig, err := ctrl.GetConfig()
-	if err != nil {
-		logger.Info("K8s config not available, K8s tools and enricher disabled", "error", err)
-		return nil
-	}
-
-	k8sClient, err := kubernetes.NewForConfig(kubeConfig)
-	if err != nil {
-		logger.Error(err, "failed to create K8s clientset")
-		return nil
-	}
-	dynClient, err := dynamic.NewForConfig(kubeConfig)
-	if err != nil {
-		logger.Error(err, "failed to create dynamic client")
-		return nil
-	}
-	cachedDisc := memory.NewMemCacheClient(k8sClient.Discovery())
-	mapper := restmapper.NewDeferredDiscoveryRESTMapper(cachedDisc)
-	return &k8sInfra{kubeConfig: kubeConfig, clientset: k8sClient, dynClient: dynClient, mapper: mapper}
-}
-
-// dsClients holds DataStorage client instances created once and shared
-// between the enricher and the custom tool registry.
-type dsClients struct {
-	ogenClient *ogenclient.Client
-	dsAdapter  *enrichment.DSAdapter
-	k8sAdapter *enrichment.K8sAdapter
-}
-
-// initDSClients creates the DataStorage adapter clients. Returns nil when
-// DataStorage URL is empty or K8s infrastructure is unavailable.
-//
-// DD-AUTH-014: When a ServiceAccount token is available (sa_token_path config
-// or default /var/run/secrets/kubernetes.io/serviceaccount/token), the ogen
-// client is configured with a Bearer token transport so that all DS API calls
-// (including ListWorkflows for the workflow validator) pass authentication.
-func initDSClients(cfg *kaconfig.Config, infra *k8sInfra, dsTokenSource *auth.TokenSource, logger logr.Logger) *dsClients {
-	if cfg.Integrations.DataStorage.URL == "" {
-		logger.Info("DataStorage URL not configured, DS adapters disabled")
-		return nil
-	}
-	if infra == nil {
-		logger.Info("K8s infrastructure unavailable, DS adapters disabled")
-		return nil
-	}
-
-	dsBase, tlsErr := buildDSBaseTransport(cfg.Integrations.DataStorage.TLS.CAFile, cfg.Integrations.DataStorage.CircuitBreaker)
-	if tlsErr != nil {
-		logger.Error(tlsErr, "failed to create TLS-aware transport for DS client",
-			"ca_file", cfg.Integrations.DataStorage.TLS.CAFile)
-		return nil
-	}
-	if cfg.Integrations.DataStorage.TLS.CAFile != "" {
-		logger.Info("DS client configured with custom TLS CA",
-			"ca_file", cfg.Integrations.DataStorage.TLS.CAFile)
-	}
-
-	const defaultDSClientTimeout = 30 * time.Second
-
-	opts := make([]ogenclient.ClientOption, 0, 1)
-	opts = append(opts, ogenclient.WithClient(&http.Client{
-		Transport: auth.NewAuthTransport(dsTokenSource, dsBase),
-		Timeout:   defaultDSClientTimeout,
-	}))
-
-	ogenClient, err := ogenclient.NewClient(cfg.Integrations.DataStorage.URL, opts...)
-	if err != nil {
-		logger.Error(err, "failed to create DataStorage ogen client", "url", cfg.Integrations.DataStorage.URL)
-		return nil
-	}
-	logger.Info("DataStorage clients initialized", "url", cfg.Integrations.DataStorage.URL)
-	return &dsClients{
-		ogenClient: ogenClient,
-		dsAdapter:  enrichment.NewDSAdapter(ogenClient),
-		k8sAdapter: newK8sAdapterWithLogger(infra, logger),
-	}
-}
-
-func newK8sAdapterWithLogger(infra *k8sInfra, logger logr.Logger) *enrichment.K8sAdapter {
-	a := enrichment.NewK8sAdapter(infra.dynClient, infra.mapper)
-	a.SetLogger(logger.WithName("k8s-adapter"))
-	kindIndex, err := k8stools.BuildKindIndex(infra.clientset.Discovery())
-	if err != nil {
-		logger.Info("failed to build kind index for K8s adapter, using empty index", "error", err)
-		kindIndex = make(map[string]schema.GroupKind)
-	}
-	a.SetKindIndex(kindIndex)
-	return a
-}
-
-// buildDSBaseTransport creates the base HTTP transport for the DataStorage
-// client. When caFile is set, uses a custom TLS transport with the specified CA;
-// otherwise falls back to the shared default transport with retry.
-// Issue #951: Wire DataStorageConfig.TLS.CAFile to DS HTTP client.
-func buildDSBaseTransport(caFile string, cbCfg types.LLMCircuitBreaker) (http.RoundTripper, error) {
-	var base http.RoundTripper
-	if caFile != "" {
-		tlsTransport, err := sharedtls.NewTLSTransport(caFile)
-		if err != nil {
-			return nil, fmt.Errorf("DS TLS transport: %w", err)
-		}
-		base = sharedtransport.NewRetryTransport(tlsTransport, sharedtransport.DefaultRetryConfig())
-	} else {
-		var err error
-		base, err = sharedtls.DefaultBaseTransportWithRetry()
-		if err != nil {
-			return nil, err
-		}
-	}
-	return sharedtransport.NewCircuitBreakerTransport(base, sharedtransport.CircuitBreakerConfig{
-		Enabled:          cbCfg.Enabled,
-		Name:             "datastorage",
-		MaxRequests:      cbCfg.MaxRequests,
-		Interval:         cbCfg.Interval,
-		Timeout:          cbCfg.Timeout,
-		FailureThreshold: cbCfg.FailureThreshold,
-		FailureRatio:     cbCfg.FailureRatio,
-	}), nil
-}
-
-// buildEnricher creates the enrichment.Enricher when DS clients are available.
-// ADR-056: attaches LabelDetector so detected_labels are populated during enrichment.
-// #704: wires RetryConfig from config for HAPI-aligned owner chain retry+fail-hard.
-func buildEnricher(cfg *kaconfig.Config, ds *dsClients, infra *k8sInfra, auditStore audit.AuditStore, logger logr.Logger) *enrichment.Enricher {
-	if ds == nil {
-		return nil
-	}
-	e := enrichment.NewEnricher(ds.k8sAdapter, ds.dsAdapter, auditStore, logger)
-	if infra != nil && infra.dynClient != nil {
-		e.WithLabelDetector(enrichment.NewLabelDetector(infra.dynClient, infra.mapper, logger.WithName("label-detector")))
-		logger.Info("label detector enabled (ADR-056)")
-	}
-	e.WithRetryConfig(enrichment.RetryConfig{
-		MaxRetries:  cfg.AI.Enrichment.MaxRetries,
-		BaseBackoff: cfg.AI.Enrichment.BaseBackoff,
-	})
-	logger.Info("enrichment retry config wired (#704)",
-		"max_retries", cfg.AI.Enrichment.MaxRetries,
-		"base_backoff", cfg.AI.Enrichment.BaseBackoff,
-	)
-	return e
-}
-
-// buildSanitizationPipeline creates the sanitization pipeline with G4 (credential scrub),
-// K8S-SECRET (JSON Secret redaction), and I1 (injection patterns) stages per DD-HAPI-019-003.
-// Returns nil when all stages are disabled.
-func buildSanitizationPipeline(cfg *kaconfig.Config, logger logr.Logger) *sanitization.Pipeline {
-	var stages []sanitization.Stage
-	if cfg.AI.Safety.Sanitization.CredentialScrubEnabled {
-		stages = append(stages, sanitization.NewCredentialSanitizer())
-	}
-	if cfg.AI.Safety.Sanitization.SecretRedactionEnabled {
-		stages = append(stages, sanitization.NewSecretSanitizer())
-	}
-	if cfg.AI.Safety.Sanitization.InjectionPatternsEnabled {
-		stages = append(stages, sanitization.NewInjectionSanitizer(nil))
-	}
-	if len(stages) == 0 {
-		logger.Info("sanitization pipeline disabled")
-		return nil
-	}
-	logger.Info("sanitization pipeline enabled", "stages", len(stages))
-	return sanitization.NewPipeline(stages...)
-}
-
-// buildAuditStore creates a BufferedDSAuditStore (DD-AUDIT-002 aligned) when audit
-// is enabled and DS is available, falling back to NopAuditStore otherwise.
-// Uses the same OpenAPIClientAdapter + BufferedAuditStore stack as every other
-// platform service. Auth transport is shared with initDSClients (same SA token)
-// to guarantee identical authentication behavior.
-func buildAuditStore(cfg *kaconfig.Config, dsTokenSource *auth.TokenSource, logger logr.Logger) (audit.AuditStore, func()) {
-	nop := func() {}
-	if !cfg.Runtime.Audit.Enabled || cfg.Integrations.DataStorage.URL == "" {
-		logger.Info("audit store disabled (nop)")
-		return audit.NopAuditStore{}, nop
-	}
-
-	auditBase, tlsErr := sharedtls.DefaultBaseTransport()
-	if tlsErr != nil {
-		logger.Error(tlsErr, "failed to create TLS-aware transport for audit store")
-		return audit.NopAuditStore{}, nop
-	}
-
-	dsClient, err := sharedaudit.NewOpenAPIClientAdapterWithTransport(
-		cfg.Integrations.DataStorage.URL, 5*time.Second, auth.NewAuthTransport(dsTokenSource, auditBase),
-	)
-	if err != nil {
-		logger.Error(err, "failed to create DS audit client, falling back to nop")
-		return audit.NopAuditStore{}, nop
-	}
-
-	var storeOpts []audit.BufferedDSAuditStoreOption
-	if cfg.Runtime.Audit.FlushIntervalSeconds > 0 {
-		storeOpts = append(storeOpts, audit.WithFlushInterval(
-			time.Duration(cfg.Runtime.Audit.FlushIntervalSeconds*float64(time.Second))))
-	}
-	if cfg.Runtime.Audit.BufferSize > 0 {
-		storeOpts = append(storeOpts, audit.WithBufferSize(cfg.Runtime.Audit.BufferSize))
-	}
-	if cfg.Runtime.Audit.BatchSize > 0 {
-		storeOpts = append(storeOpts, audit.WithBatchSize(cfg.Runtime.Audit.BatchSize))
-	}
-
-	store, err := audit.NewBufferedDSAuditStore(dsClient, logger, storeOpts...)
-	if err != nil {
-		logger.Error(err, "failed to create buffered audit store, falling back to nop")
-		return audit.NopAuditStore{}, nop
-	}
-	logger.Info("audit store enabled (buffered, DD-AUDIT-002 aligned)",
-		"ds_url", cfg.Integrations.DataStorage.URL)
-	return store, func() {
-		if closeErr := store.Close(); closeErr != nil {
-			logger.Error(closeErr, "audit store close error")
-		}
-	}
-}
-
-// buildSummarizer creates a tool output summarizer when the threshold is positive.
-// When MaxToolOutputSize is configured, it enables pre-truncation to prevent
-// the summarizer's own LLM call from exceeding context window limits (#752).
-func buildSummarizer(llmClient llm.Client, cfg *kaconfig.Config, logger logr.Logger) *summarizer.Summarizer {
-	if cfg.AI.Summarizer.Threshold <= 0 {
-		logger.Info("summarizer disabled (threshold <= 0)")
-		return nil
-	}
-	if cfg.AI.Summarizer.MaxToolOutputSize > 0 {
-		logger.Info("summarizer enabled with pre-truncation",
-			"threshold", cfg.AI.Summarizer.Threshold,
-			"max_tool_output_size", cfg.AI.Summarizer.MaxToolOutputSize)
-		return summarizer.NewWithMaxInput(llmClient, cfg.AI.Summarizer.Threshold, cfg.AI.Summarizer.MaxToolOutputSize)
-	}
-	logger.Info("summarizer enabled", "threshold", cfg.AI.Summarizer.Threshold)
-	return summarizer.New(llmClient, cfg.AI.Summarizer.Threshold)
-}
-
-// buildAnomalyDetector creates the I7 anomaly detector from config thresholds.
-func buildAnomalyDetector(cfg *kaconfig.Config, logger logr.Logger) *investigator.AnomalyDetector {
-	ac := investigator.AnomalyConfig{
-		MaxToolCallsPerTool: cfg.AI.Safety.Anomaly.MaxToolCallsPerTool,
-		MaxTotalToolCalls:   cfg.AI.Safety.Anomaly.MaxTotalToolCalls,
-		MaxRepeatedFailures: cfg.AI.Safety.Anomaly.MaxRepeatedFailures,
-		ExemptPrefixes:      cfg.AI.Safety.Anomaly.ExemptPrefixes,
-	}
-	logger.Info("anomaly detector enabled",
-		"maxToolCallsPerTool", ac.MaxToolCallsPerTool,
-		"maxTotalToolCalls", ac.MaxTotalToolCalls,
-		"maxRepeatedFailures", ac.MaxRepeatedFailures,
-		"exemptPrefixes", ac.ExemptPrefixes,
-	)
-	return investigator.NewAnomalyDetector(ac, nil)
-}
-
-// buildToolRegistry creates and populates the tool registry with all available tool sets.
-func buildToolRegistry(cfg *kaconfig.Config, logger logr.Logger, infra *k8sInfra, ds *dsClients, auditStore audit.AuditStore) *registry.Registry {
-	reg := registry.New()
-
-	if infra != nil {
-		registerK8sTools(reg, infra, logger, auditStore)
-	}
-
-	if cfg.Integrations.Tools.Prometheus.URL != "" {
-		promCfg := promtools.ClientConfig{
-			URL:       cfg.Integrations.Tools.Prometheus.URL,
-			Timeout:   cfg.Integrations.Tools.Prometheus.Timeout,
-			SizeLimit: cfg.Integrations.Tools.Prometheus.SizeLimit,
-		}
-		if cfg.Integrations.Tools.Prometheus.TLSCaFile != "" {
-			promBase, promTLSErr := sharedtls.NewTLSTransport(cfg.Integrations.Tools.Prometheus.TLSCaFile)
-			if promTLSErr != nil {
-				logger.Error(promTLSErr, "failed to create Prometheus TLS transport", "ca_file", cfg.Integrations.Tools.Prometheus.TLSCaFile)
-			} else {
-				promCfg.Transport = auth.NewAuthTransport(auth.NewDefaultTokenSource(), promBase)
-				logger.Info("Prometheus client configured with TLS + SA bearer auth", "ca_file", cfg.Integrations.Tools.Prometheus.TLSCaFile)
-			}
-		}
-		promClient, promErr := promtools.NewClient(promCfg)
-		if promErr != nil {
-			logger.Error(promErr, "failed to create Prometheus client")
-		} else {
-			for _, t := range promtools.NewAllTools(promClient) {
-				reg.Register(t)
-			}
-			logger.Info("registered Prometheus tools", "count", len(promtools.AllToolNames))
-		}
-	}
-
-	if cfg.Integrations.Tools.Alertmanager.URL != "" {
-		amCfg := amtools.ClientConfig{
-			URL:       cfg.Integrations.Tools.Alertmanager.URL,
-			Timeout:   cfg.Integrations.Tools.Alertmanager.Timeout,
-			SizeLimit: cfg.Integrations.Tools.Alertmanager.SizeLimit,
-		}
-		if cfg.Integrations.Tools.Alertmanager.TLSCaFile != "" {
-			amBase, amTLSErr := sharedtls.NewTLSTransport(cfg.Integrations.Tools.Alertmanager.TLSCaFile)
-			if amTLSErr != nil {
-				logger.Error(amTLSErr, "failed to create Alertmanager TLS transport", "ca_file", cfg.Integrations.Tools.Alertmanager.TLSCaFile)
-			} else {
-				amCfg.Transport = auth.NewAuthTransport(auth.NewDefaultTokenSource(), amBase)
-				logger.Info("Alertmanager client configured with TLS + SA bearer auth", "ca_file", cfg.Integrations.Tools.Alertmanager.TLSCaFile)
-			}
-		}
-		amClient, amErr := amtools.NewClient(amCfg)
-		if amErr != nil {
-			logger.Error(amErr, "failed to create Alertmanager client")
-		} else {
-			for _, t := range amtools.NewAllTools(amClient) {
-				reg.Register(t)
-			}
-			logger.Info("registered Alertmanager tools", "count", len(amtools.AllToolNames))
-		}
-	}
-
-	if ds != nil {
-		custom.RegisterAll(reg, ds.ogenClient, ds.dsAdapter, ds.k8sAdapter, logger)
-		logger.Info("registered custom tools", "count", len(custom.AllToolNames))
-	}
-
-	reg.Register(investigation.NewTodoWriteTool())
-	logger.Info("registered TodoWrite tool")
-
-	logger.Info("tool registry ready", "total_tools", len(reg.All()))
-	return reg
-}
-
-// resolveAlignmentCheckConfig returns the effective AlignmentCheckConfig.
-// When fleet mode is active and defines its own alignment check, the fleet
-// override takes precedence over the global ai.alignmentCheck. This allows
-// operators to enforce cross-model shadow evaluation specifically for
-// multi-cluster investigations where prompt injection risk is higher.
-func resolveAlignmentCheckConfig(cfg *kaconfig.Config) kaconfig.AlignmentCheckConfig {
-	fleetCfg := cfg.Integrations.Fleet
-	if fleetCfg.GatewayType != "" && fleetCfg.Endpoint != "" && fleetCfg.AlignmentCheck != nil {
-		return *fleetCfg.AlignmentCheck
-	}
-	return cfg.AI.AlignmentCheck
-}
-
-// registerFleetTools connects to the MCP Gateway, creates a GatewayDiscoverer
-// for the configured gateway type, pre-scopes tools for the target cluster,
-// and registers list_clusters + list_tools_for_cluster for LLM-driven discovery.
-// Returns the fleet client (must be closed on shutdown) and the registered tool names,
-// or nil if fleet is disabled.
-//
-// Authority: ADR-068 decision #11
-func registerFleetTools(ctx context.Context, cfg *kaconfig.Config, reg *registry.Registry, logger logr.Logger) (*fleetclient.ResilientClient, []string) {
-	gatewayType := cfg.Integrations.Fleet.GatewayType
-	endpoint := cfg.Integrations.Fleet.Endpoint
-	if gatewayType == "" || endpoint == "" {
-		return nil, nil
-	}
-
-	fleetLog := logger.WithName("fleet")
-	fleetLog.Info("connecting to MCP Gateway for fleet tool discovery",
-		"endpoint", endpoint, "gatewayType", gatewayType)
-
-	var opts []fleetclient.Option
-	if cfg.Integrations.Fleet.OAuth2.Enabled {
-		basePath := "/etc/kubernautagent/fleet-oauth2"
-		if cfg.Integrations.Fleet.OAuth2.CredentialsSecretRef != "" {
-			basePath = "/etc/kubernautagent/" + cfg.Integrations.Fleet.OAuth2.CredentialsSecretRef
-		}
-		reloadCfg := fleetclient.ReloadableOAuth2Config{
-			TokenURL:         cfg.Integrations.Fleet.OAuth2.TokenURL,
-			ClientIDPath:     basePath + "/client-id",
-			ClientSecretPath: basePath + "/client-secret",
-			Scopes:           fleetclient.DefaultFleetScopes(cfg.Integrations.Fleet.OAuth2.Scopes),
-			TokenTimeout:     10 * time.Second,
-		}
-		opts = append(opts, fleetclient.WithReloadableOAuth2Transport(reloadCfg, fleetLog))
-		fleetLog.Info("fleet OAuth2 authentication configured (hot-reloadable)",
-			"tokenURL", cfg.Integrations.Fleet.OAuth2.TokenURL,
-			"secretPath", basePath)
-	}
-
-	resilienceCfg := fleetclient.DefaultResilienceConfig()
-	resilientClient, err := fleetclient.NewResilient(ctx, endpoint, resilienceCfg, fleetLog, opts...)
-	if err != nil {
-		fleetLog.Error(err, "failed to connect to MCP Gateway — fleet tools unavailable")
-		return nil, nil
-	}
-
-	session := resilientClient.Session()
-
-	discoverer, err := fleetclient.NewDiscoverer(fleetregistry.MCPGatewayType(gatewayType), session)
-	if err != nil {
-		fleetLog.Error(err, "failed to create GatewayDiscoverer", "gatewayType", gatewayType)
-		_ = resilientClient.Close()
-		return nil, nil
-	}
-
-	listClustersTool := fleetclient.NewListClustersTool(discoverer)
-	listToolsTool := fleetclient.NewListToolsForClusterTool(discoverer, reg, session)
-
-	reg.Register(listClustersTool)
-	reg.Register(listToolsTool)
-
-	toolNames := make([]string, 0, 2)
-	toolNames = append(toolNames, listClustersTool.Name(), listToolsTool.Name())
-
-	fleetLog.Info("registered fleet discovery tools",
-		"tools", toolNames, "endpoint", endpoint, "gatewayType", gatewayType)
-	return resilientClient, toolNames
-}
-
-func registerK8sTools(reg *registry.Registry, infra *k8sInfra, logger logr.Logger, auditStore audit.AuditStore) {
-	kindIndex, err := k8stools.BuildKindIndex(infra.clientset.Discovery())
-	if err != nil {
-		logger.Info("failed to build kind index, using empty index", "error", err)
-		kindIndex = make(map[string]schema.GroupKind)
-	}
-	resolver := k8stools.NewDynamicResolver(infra.dynClient, infra.mapper, kindIndex, logger.WithName("k8s-resolver"),
-		k8stools.WithSecretAccessObserver(newSecretAccessObserver(auditStore, logger.WithName("secret-access-audit"))))
-
-	for _, t := range k8stools.NewAllTools(infra.clientset, resolver) {
-		reg.Register(t)
-	}
-	logger.Info("registered K8s tools", "count", len(k8stools.AllToolNames))
-
-	reg.Register(logtools.NewFetchPodLogsTool(infra.clientset))
-	logger.Info("registered fetch_pod_logs tool")
-
-	mc, mcErr := metricsclient.NewForConfig(infra.kubeConfig)
-	if mcErr != nil {
-		logger.Error(mcErr, "failed to create metrics client, metrics tools will not be registered")
-	} else {
-		for _, t := range k8stools.NewMetricsTools(k8stools.NewMetricsClient(mc)) {
-			reg.Register(t)
-		}
-		logger.Info("registered metrics tools", "count", len(k8stools.MetricsToolNames))
-	}
-
-	npc := k8stools.NewNodeProxyClient(infra.clientset)
-	for _, t := range k8stools.NewNodeProxyTools(npc, 30000) {
-		reg.Register(t)
-	}
-	logger.Info("registered node proxy tools", "count", len(k8stools.NodeProxyToolNames))
-}
-
-// newSecretAccessObserver builds the k8stools.SecretAccessObserver wired into
-// the K8s resource resolver (GAP-13, Issue #1505). It is the detective
-// control compensating for KubernautAgent's intentionally broad read RBAC on
-// Secrets (see docs/services/stateless/kubernaut-agent/security-configuration.md):
-// every Secret Get/List — success or failure — becomes an independently
-// queryable aiagent.secret.accessed audit event, correlated to the
-// investigation via the correlationID set on ctx by
-// session.Manager.launchInvestigation.
-func newSecretAccessObserver(auditStore audit.AuditStore, logger logr.Logger) func(ctx context.Context, verb, name, namespace string, err error) {
-	return func(ctx context.Context, verb, name, namespace string, accessErr error) {
-		if auditStore == nil {
-			return
-		}
-		correlationID, _ := audit.CorrelationIDFromContext(ctx)
-
-		event := audit.NewEvent(audit.EventTypeSecretAccessed, correlationID)
-		event.EventAction = audit.ActionSecretAccessed
-		event.EventOutcome = audit.OutcomeSuccess
-		event.Data["verb"] = verb
-		if namespace != "" {
-			event.Data["namespace"] = namespace
-		}
-		if name != "" {
-			event.Data["secret_name"] = name
-		}
-		if accessErr != nil {
-			event.EventOutcome = audit.OutcomeFailure
-			event.Data["outcome_detail"] = accessErr.Error()
-		}
-		audit.StoreBestEffort(ctx, auditStore, event, logger)
-	}
-}
-
-// dsCatalogFetcher implements investigator.CatalogFetcher by querying
-// DataStorage on every call. This removes the boot-time blocking fetch
-// that caused #665 (CrashLoopBackOff when the catalog was not yet seeded).
-//
-// Per DD-HAPI-002 (v1.1+), KA is the sole workflow validator. The catalog
-// is fetched per-request so KA always validates against the current catalog
-// without needing a restart when workflows are added/removed.
-type dsCatalogFetcher struct {
-	ds     *dsClients
-	logger logr.Logger
-}
-
-func newDSCatalogFetcher(ds *dsClients, logger logr.Logger) *dsCatalogFetcher {
-	return &dsCatalogFetcher{ds: ds, logger: logger}
-}
-
-func (f *dsCatalogFetcher) FetchValidator(ctx context.Context) (*parser.Validator, error) {
-	fetchCtx, cancel := context.WithTimeout(ctx, 15*time.Second)
-	defer cancel()
-
-	resp, err := f.ds.ogenClient.ListWorkflows(fetchCtx, ogenclient.ListWorkflowsParams{})
-	if err != nil {
-		return nil, fmt.Errorf("ListWorkflows call failed: %w", err)
-	}
-
-	wlr, ok := resp.(*ogenclient.WorkflowListResponse)
-	if !ok {
-		return nil, fmt.Errorf("unexpected ListWorkflows response type %T", resp)
-	}
-
-	ids := make([]string, 0, len(wlr.Workflows))
-	for _, w := range wlr.Workflows {
-		if w.WorkflowId.Set {
-			ids = append(ids, w.WorkflowId.Value.String())
-		}
-	}
-	if len(ids) == 0 {
-		return nil, fmt.Errorf("workflow catalog returned 0 workflows")
-	}
-
-	validator := parser.NewValidator(ids)
-	schemaParser := dsschema.NewParser()
-	for _, w := range wlr.Workflows {
-		if !w.WorkflowId.Set {
-			continue
-		}
-		wfID := w.WorkflowId.Value.String()
-		meta := parser.WorkflowMeta{
-			ExecutionEngine: w.ExecutionEngine,
-			Version:         w.Version,
-			Component:       append([]string(nil), w.Labels.GetComponent()...),
-		}
-		if w.ExecutionBundle.Set {
-			meta.ExecutionBundle = w.ExecutionBundle.Value
-		}
-		if w.ExecutionBundleDigest.Set {
-			meta.ExecutionBundleDigest = w.ExecutionBundleDigest.Value
-		}
-		if w.ServiceAccountName.Set {
-			meta.ServiceAccountName = w.ServiceAccountName.Value
-		}
-		if w.Content != "" {
-			parsed, err := schemaParser.Parse(w.Content)
-			if err != nil {
-				f.logger.Error(err, "failed to parse workflow schema Content, parameter validation will strip all LLM params (fail-closed)",
-					"workflow_id", wfID)
-			} else {
-				meta.Parameters = parsed.Parameters
-			}
-		}
-		validator.SetWorkflowMeta(wfID, meta)
-	}
-
-	f.logger.Info("workflow catalog fetched (DD-HAPI-002: per-request validation)",
-		"allowed_workflows", len(ids))
-	return validator, nil
-}
-
-// newAuthMiddleware creates the DD-AUTH-014 auth middleware using the shared k8sInfra clientset.
-// DD-AUTH-MCP-001 v2.0: When JWT providers are configured, wraps K8sAuthenticator
-// in a CompositeAuthenticator for Pattern A + Pattern B coexistence.
-// Returns a cleanup function that stops JWKS background goroutines (nil if no JWT).
-func newAuthMiddleware(infra *k8sInfra, interactiveCfg kaconfig.InteractiveConfig, logger logr.Logger) (*auth.Middleware, func()) {
-	if infra == nil || infra.clientset == nil {
-		logger.Info("K8s infrastructure not available, auth middleware disabled")
-		return nil, nil
-	}
-
-	k8sAuth := auth.NewK8sAuthenticator(infra.clientset)
-	authorizer := auth.NewK8sAuthorizer(infra.clientset)
-	namespace := detectNamespace()
-
-	var authenticator auth.Authenticator = k8sAuth
-	var cleanup func()
-
-	if interactiveCfg.Enabled && len(interactiveCfg.JWTProviders) > 0 {
-		entries := make([]auth.JWTProviderEntry, len(interactiveCfg.JWTProviders))
-		for i, p := range interactiveCfg.JWTProviders {
-			entries[i] = auth.JWTProviderEntry{
-				Issuer:        p.Issuer,
-				JWKSURL:       p.JWKSURL,
-				Audience:      p.Audience,
-				UsernameClaim: p.ClaimMappings.Username,
-				GroupsClaim:   p.ClaimMappings.Groups,
-				TLSCAFile:     p.TLSCaFile,
-			}
-		}
-
-		for _, e := range entries {
-			if strings.HasPrefix(e.JWKSURL, "http://") {
-				logger.Info("WARNING: JWKS URL uses plain HTTP — vulnerable to MITM in production; enforce HTTPS via kubernaut-operator admission webhook (kubernaut-operator#46)",
-					"provider", e.Issuer, "jwksURL", e.JWKSURL)
-			}
-		}
-
-		jwtAuth, err := auth.NewJWTAuthenticator(entries, logger.WithName("jwt-auth"))
-		if err != nil {
-			logger.Error(err, "failed to create JWTAuthenticator; Pattern B disabled, Pattern A active")
-		} else {
-			authenticator = auth.NewCompositeAuthenticator(jwtAuth, k8sAuth)
-			cleanup = func() {
-				jwtAuth.Close()
-				logger.Info("JWTAuthenticator JWKS caches stopped")
-			}
-			logger.Info("CompositeAuthenticator enabled (Pattern A + Pattern B)",
-				"jwtProviders", len(entries),
-			)
-		}
-	}
-
-	mw := auth.NewMiddleware(authenticator, authorizer, auth.MiddlewareConfig{
-		Namespace:    namespace,
-		Resource:     "services",
-		ResourceName: "kubernaut-agent",
-		Verb:         "create",
-	}, logger)
-	return mw, cleanup
-}
-
-// buildMCPHandler constructs the fully-wired MCP interactive handler with all
-// tools registered. Returns nil if prerequisites are missing (K8s infra or DS).
-// PR6a: Production wiring for MCP interactive mode (BR-INTERACTIVE-001..008).
-// mcpHandlerParams groups the dependencies needed to build the MCP
-// interactive-mode HTTP handler. Extracted per AGENTS.md's 8+-param
-// Options-pattern rule.
-// apiRoutesParams groups the dependencies needed to register the /api/v1
-// route tree (body-size limiting, throttling, metrics, rate limiting, auth,
-// the optional MCP interactive endpoint, and the ogen-generated REST API).
-type apiRoutesParams struct {
-	cfg                  *kaconfig.Config
-	infra                *k8sInfra
-	ds                   *dsClients
-	inv                  *investigator.Investigator
-	enricher             *enrichment.Enricher
-	mgr                  *session.Manager
-	agentMetrics         *kametrics.Metrics
-	instrumentedAudit    audit.AuditStore
-	ogenSrv              http.Handler
-	eventEmitter         *karbac.EventEmitter
-	interactiveReadiness *karbac.InteractiveReadiness
-	apiRateLimiter       *kaserver.RateLimiter
-	maxRequestBodySize   int64
-	logger               logr.Logger
-}
-
-// registerAPIRoutes mounts the /api/v1 route tree onto r: request body-size
-// limiting, optional concurrency throttling, HTTP metrics, API rate limiting,
-// auth middleware (DD-AUTH-014), the optional MCP interactive endpoint, and
-// the ogen-generated REST API as the catch-all. Returns the MCP session
-// drainer (nil when interactive mode is disabled or handler construction
-// failed) and the auth middleware's JWKS cache cleanup func (nil when auth
-// is disabled), both intended to be handled by the caller after this
-// function returns.
-func registerAPIRoutes(r chi.Router, ctx context.Context, p apiRoutesParams) (*mcpkg.SessionDrainer, func()) {
-	var sessionDrainer *mcpkg.SessionDrainer
-	var authCleanupRef func()
-
-	r.Route("/api/v1", func(r chi.Router) {
-		r.Use(func(next http.Handler) http.Handler {
-			return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-				req.Body = http.MaxBytesReader(w, req.Body, p.maxRequestBodySize)
-				next.ServeHTTP(w, req)
-			})
-		})
-		if p.cfg.Runtime.Server.MaxConcurrentRequests > 0 {
-			p.logger.Info("concurrency throttling enabled",
-				"max_concurrent_requests", p.cfg.Runtime.Server.MaxConcurrentRequests)
-			r.Use(chimiddleware.Throttle(p.cfg.Runtime.Server.MaxConcurrentRequests))
-		}
-		r.Use(kaserver.HTTPMetricsMiddleware(p.agentMetrics))
-		r.Use(p.apiRateLimiter.Middleware)
-
-		authMw, authCleanup := newAuthMiddleware(p.infra, p.cfg.Interactive, p.logger)
-		if authCleanup != nil {
-			// authCleanupRef is returned to the caller and deferred in main()'s
-			// scope (not here — this is a chi route-setup closure that returns
-			// immediately).
-			authCleanupRef = authCleanup
-		}
-		if authMw != nil {
-			r.Use(func(next http.Handler) http.Handler {
-				return kaserver.AuditAuthMiddleware(authMw.Handler(next), p.instrumentedAudit, p.logger)
-			})
-			p.logger.Info("auth middleware enabled (DD-AUTH-014)",
-				"resource", "services",
-				"resourceName", "kubernaut-agent",
-				"verb", "create",
-			)
-		} else {
-			p.logger.Info("auth middleware DISABLED (no in-cluster K8s config)")
-		}
-
-		if p.cfg.Interactive.Enabled {
-			var mcpHandler http.Handler
-			mcpHandler, sessionDrainer = buildMCPHandler(ctx, mcpHandlerParams{
-				cfg: p.cfg, infra: p.infra, ds: p.ds, inv: p.inv, enricher: p.enricher,
-				autoMgr: p.mgr, authMw: authMw, agentMetrics: p.agentMetrics,
-				auditStore: p.instrumentedAudit, logger: p.logger,
-			})
-			if mcpHandler != nil {
-				// SEC-02: Per-user rate limiting for MCP interactive endpoint.
-				userRL := kaserver.NewUserRateLimiter(
-					kaserver.DefaultUserRateLimitConfig(p.cfg.Interactive.RateLimitPerUser),
-					p.agentMetrics.HTTPRateLimitedTotal,
-				)
-				defer userRL.Stop()
-
-				r.Route("/mcp", func(mcpRouter chi.Router) {
-					mcpRouter.Use(userRL.Middleware)
-					mcpRouter.Handle("/", kaserver.SSEHeadersMiddleware(mcpHandler))
-					mcpRouter.Handle("/*", kaserver.SSEHeadersMiddleware(mcpHandler))
-				})
-				p.interactiveReadiness.SetEnabled()
-				if p.eventEmitter != nil {
-					p.eventEmitter.EmitInteractiveEnabled()
-				}
-				p.logger.Info("MCP interactive route mounted",
-					"path", "/api/v1/mcp",
-					"rateLimitPerUser", p.cfg.Interactive.RateLimitPerUser,
-				)
-			} else {
-				p.interactiveReadiness.SetSoftDisabled("handler construction failed (check preceding errors)")
-				if p.eventEmitter != nil {
-					p.eventEmitter.EmitInteractiveSoftDisabled("MCP handler construction failed")
-				}
-				p.logger.Error(nil, "MCP interactive mode enabled but handler construction failed (check preceding errors)")
-			}
-		}
-
-		r.Handle("/*", kaserver.SSEHeadersMiddleware(p.ogenSrv))
-	})
-
-	return sessionDrainer, authCleanupRef
-}
-
-type mcpHandlerParams struct {
-	cfg          *kaconfig.Config
-	infra        *k8sInfra
-	ds           *dsClients
-	inv          *investigator.Investigator
-	enricher     *enrichment.Enricher
-	autoMgr      *session.Manager
-	authMw       *auth.Middleware
-	agentMetrics *kametrics.Metrics
-	auditStore   audit.AuditStore
-	logger       logr.Logger
-}
-
-func buildMCPHandler(ctx context.Context, p mcpHandlerParams) (http.Handler, *mcpkg.SessionDrainer) {
-	cfg, infra, ds, inv, enricher, autoMgr, authMw, agentMetrics, auditStore, logger :=
-		p.cfg, p.infra, p.ds, p.inv, p.enricher, p.autoMgr, p.authMw, p.agentMetrics, p.auditStore, p.logger
-
-	if infra == nil || infra.kubeConfig == nil {
-		logger.Error(nil, "MCP interactive mode: K8s infrastructure unavailable")
-		return nil, nil
-	}
-	if authMw == nil {
-		logger.Error(nil, "MCP interactive mode: auth middleware unavailable (DD-AUTH-MCP-001)")
-		return nil, nil
-	}
-	// SEC-05: Investigator is required for the core investigate tool.
-	if inv == nil {
-		logger.Error(nil, "MCP interactive mode: investigator unavailable")
-		return nil, nil
-	}
-
-	// SEC-07: Build controller-runtime client with MCP-specific timeouts.
-	// Scheme includes remediationv1 for RR existence validation (HARM-004)
-	// and future NL signal intake (#714).
-	mcpScheme := k8sruntime.NewScheme()
-	utilruntime.Must(clientgoscheme.AddToScheme(mcpScheme))
-	utilruntime.Must(remediationv1.AddToScheme(mcpScheme))
-
-	mcpRestConfig := *infra.kubeConfig
-	mcpRestConfig.Timeout = 10 * time.Second
-	mcpRestConfig.QPS = 20
-	mcpRestConfig.Burst = 40
-
-	ctrlCli, err := ctrlclient.New(&mcpRestConfig, ctrlclient.Options{Scheme: mcpScheme})
-	if err != nil {
-		logger.Error(err, "MCP interactive mode: failed to create controller-runtime client")
-		return nil, nil
-	}
-
-	namespace := detectNamespace()
-
-	// emitDisconnectAudit emits interactive.completed for non-tool session endings
-	// (disconnect, inactivity timeout, TTL expiry). M1: ensures all session-ending
-	// paths produce an audit trail, not just action=complete/cancel through InvestigateTool.
-	emitDisconnectAudit := func(sessionID, correlationID, reason string) {
-		event := audit.NewEvent(audit.EventTypeInteractiveCompleted, correlationID,
-			audit.WithSessionID(sessionID),
-		)
-		event.EventAction = audit.ActionInteractiveCompleted
-		event.EventOutcome = audit.OutcomeSuccess
-		event.Data["reason"] = reason
-		audit.StoreBestEffort(context.Background(), auditStore, event, logger.WithName("mcp-audit"))
-	}
-
-	// Session management via K8s Leases (single-driver guarantee).
-	leaseOpts := []mcpkg.LeaseOption{
-		mcpkg.WithSessionTTL(cfg.Interactive.SessionTTL),
-		mcpkg.WithInactivityTimeout(cfg.Interactive.InactivityTimeout),
-		mcpkg.WithMaxConcurrentSessions(cfg.Interactive.MaxConcurrentSessions),
-		mcpkg.WithSessionExpiredCallback(func(sessionID, rrID, reason string) {
-			// #1438: Emit terminal event BEFORE completing the HTTP session so
-			// EventLogBridge can forward it to AF before the channel closes.
-			autoMgr.EmitSessionEndedByRR(rrID, reason)
-			mcptools.CompleteHTTPSession(autoMgr, rrID, nil, logger, reason)
-			emitDisconnectAudit(sessionID, rrID, reason)
-			agentMetrics.RecordInteractiveSessionEnded()
-		}),
-	}
-	leaseMgr := mcpkg.NewLeaseSessionManagerConcrete(ctrlCli, namespace, logger, leaseOpts...)
-
-	if n := leaseMgr.ReconcileOrphanedLeases(context.Background()); n > 0 {
-		logger.Info("startup: reclaimed orphaned interactive Leases", "count", n)
-	}
-
-	// Context reconstruction from DS audit events (best-effort).
-	var recon mcpkg.ContextReconstructor
-	if ds != nil {
-		recon = mcpkg.NewDSContextReconstructor(ds.ogenClient, 10*time.Second, logger)
-	} else {
-		recon = &noopReconstructor{}
-		logger.Info("MCP interactive mode: DS unavailable — context reconstruction disabled")
-	}
-
-	// DelegatingEventStore: bridges MCP SDK session lifecycle to our disconnect
-	// handler. Wraps SDK's MemoryEventStore for stream resumption support.
-	eventStore := mcpkg.NewDelegatingEventStore()
-
-	// TimeoutManager: fires onExpire when a session goes inactive (SEC-04, HARM-03/04).
-	timeoutMgr := mcpkg.NewTimeoutManager(
-		cfg.Interactive.InactivityTimeout,
-		[]time.Duration{cfg.Interactive.InactivityTimeout - 2*time.Minute, cfg.Interactive.InactivityTimeout - 30*time.Second},
-		func(sessionID string) {
-			logger.Info("interactive session expired due to inactivity",
-				"session_id", sessionID)
-			// Snapshot correlationID before Release deletes the entry.
-			rrID, _ := leaseMgr.GetSessionInfo(sessionID)
-			// #1438: Emit terminal event BEFORE closing the HTTP session so the
-			// EventLogBridge can forward it to AF before the channel closes.
-			autoMgr.EmitSessionEndedByRR(rrID, "inactivity_timeout")
-			if err := leaseMgr.Release(sessionID, "inactivity_timeout"); err != nil {
-				logger.Error(err, "failed to release expired session",
-					"session_id", sessionID)
-				return
-			}
-			// KA-CRIT-2: Resolve the HTTP session so AA stops polling user_driving.
-			mcptools.CompleteHTTPSession(autoMgr, rrID, nil, logger, "inactivity_timeout")
-			emitDisconnectAudit(sessionID, rrID, "inactivity_timeout")
-			// T1-4: Decrement gauge on timeout expiry to prevent drift.
-			agentMetrics.RecordInteractiveSessionEnded()
-		},
-	)
-
-	// ReconstructionSpawner: rebuilds context and spawns autonomous investigation
-	// after an interactive session ends (INT-06, BR-INTERACTIVE-008).
-	reconRunner := mcpadapters.NewReconRunnerAdapter(inv)
-	reconSpawner := mcpkg.NewReconstructionSpawner(reconRunner, recon, logger)
-
-	// GracefulSessionClosedHandler: processes MCP disconnect events with a
-	// configurable grace period before releasing the interactive lease.
-	// BR-INTERACTIVE-001: decouples MCP transport lifecycle from lease lifecycle.
-	// If a client reconnects during the grace period, the lease is preserved.
-	disconnectGracePeriod := cfg.Interactive.DisconnectGracePeriod
-	if disconnectGracePeriod <= 0 {
-		disconnectGracePeriod = 60 * time.Second
-	}
-	disconnectHandler := mcpkg.NewGracefulSessionClosedHandler(eventStore, func(mcpSessionID string) {
-		interactiveSessionID, ok := eventStore.LookupInteractiveSession(mcpSessionID)
-		if !ok {
-			logger.V(1).Info("MCP session closed without interactive mapping (autonomous or already released)",
-				"mcp_session_id", mcpSessionID)
-			return
-		}
-
-		eventStore.DeleteMCPSession(mcpSessionID)
-
-		timeoutMgr.StopTracking(interactiveSessionID)
-
-		// T1-1: Snapshot session info BEFORE Release deletes the entry.
-		rrID, signalMeta := leaseMgr.GetSessionInfo(interactiveSessionID)
-
-		// #1438: Emit terminal event BEFORE closing the HTTP session so the
-		// EventLogBridge can forward it to AF before the channel closes.
-		autoMgr.EmitSessionEndedByRR(rrID, "disconnect")
-
-		if err := leaseMgr.Release(interactiveSessionID, "disconnect"); err != nil {
-			logger.Info("failed to release disconnected session",
-				"session_id", interactiveSessionID,
-				"error", err.Error())
-			return
-		}
-
-		// KA-CRIT-2: Resolve the HTTP session so AA stops polling user_driving.
-		mcptools.CompleteHTTPSession(autoMgr, rrID, nil, logger, "disconnect")
-
-		emitDisconnectAudit(interactiveSessionID, rrID, "disconnect")
-
-		// T1-4: Decrement gauge on disconnect to prevent drift.
-		agentMetrics.RecordInteractiveSessionEnded()
-
-		// Spawn reconstruction in background (best-effort, BR-INTERACTIVE-008).
-		go func() {
-			reconCtx, reconCancel := context.WithTimeout(context.Background(), 5*time.Minute)
-			defer reconCancel()
-			if err := reconSpawner.SpawnReconstruct(reconCtx, &mcpkg.ReconstructionContext{
-				CorrelationID: rrID,
-				SessionID:     interactiveSessionID,
-				SignalMeta:    signalMeta,
-			}); err != nil {
-				logger.Error(err, "background reconstruction failed",
-					"correlationID", rrID, "sessionID", interactiveSessionID)
-			}
-		}()
-	}, disconnectGracePeriod, logger)
-
-	// Wire reconnect callback: when Takeover detects a same-user reconnect,
-	// cancel the pending graceful release (BR-INTERACTIVE-001, #1442).
-	leaseMgr.SetReconnectCallback(func(sessionID string) {
-		if disconnectHandler.CancelPendingRelease(sessionID) {
-			logger.Info("pending disconnect release cancelled (client reconnected)",
-				"session_id", sessionID)
-		}
-	})
-
-	// Start disconnect handler goroutine.
-	go disconnectHandler.Run(ctx)
-
-	// Build the InvestigatorRunner adapter.
-	investigatorRunner := mcpadapters.NewInvestigatorRunnerAdapter(inv)
-
-	// SEC-HIGH-01: Per-session message rate limiter (maxMessageSize = 64KB).
-	sessionRateLimiter := mcpkg.NewSessionRateLimiter(cfg.Interactive.RateLimitPerUser, 64*1024)
-
-	// UX-01/02: Session notifier delivers timeout warnings to MCP clients.
-	sessionNotifier := mcpkg.NewSessionNotifier()
-
-	// HARM-004: Validate RR existence before creating interactive Leases.
-	rrChecker := mcptools.NewK8sRRExistenceChecker(ctrlCli, namespace)
-
-	// Signal context resolver: reads the SignalContext stored on the session
-	// from the original AA IncidentRequest payload. Falls back to reading
-	// the RR CRD for sessions without stored signal (e.g. interactive sessions
-	// started directly via MCP without an AA payload).
-	signalResolver := mcpadapters.NewSessionSignalContextResolver(autoMgr, ctrlCli, namespace)
-
-	// Build the WorkflowCatalog adapter (shared between InvestigateTool and SelectWorkflowTool).
-	wfQuerier := wfclient.NewOgenWorkflowQuerier(ds.ogenClient)
-	catalogAdapter := mcpadapters.NewWorkflowCatalogAdapter(wfQuerier)
-
-	// Build the InvestigateTool with optional dependencies.
-	investigateOpts := []mcptools.InvestigateOption{
-		mcptools.WithToolMetrics(agentMetrics),
-		mcptools.WithRateLimiter(sessionRateLimiter),
-		mcptools.WithTimeoutTracker(timeoutMgr),
-		mcptools.WithNotifyFunc(sessionNotifier.Notify),
-		mcptools.WithRRExistenceChecker(rrChecker),
-		mcptools.WithHTTPCompleter(autoMgr),
-		mcptools.WithAuditStore(auditStore, logger.WithName("mcp-audit")),
-		mcptools.WithSignalContextResolver(signalResolver),
-		mcptools.WithWorkflowCatalog(catalogAdapter),
-	}
-	investigateTool := mcptools.NewInvestigateTool(leaseMgr, investigatorRunner, recon, autoMgr, investigateOpts...)
-
-	// Build SelectWorkflowTool (reuses the same catalogAdapter).
-	// #1012: enrichment is now internalized into select_workflow via WithEnrichmentRunner.
-	swOpts := []mcptools.SelectWorkflowOption{
-		mcptools.WithLogger(logger.WithName("select-workflow")),
-		mcptools.WithHTTPSessionCompleter(autoMgr),
-		mcptools.WithMutexProvider(investigateTool),
-	}
-	if enricher != nil {
-		swOpts = append(swOpts, mcptools.WithEnrichmentRunner(enricher))
-	}
-	selectWfTool := mcptools.NewSelectWorkflowTool(catalogAdapter, leaseMgr, swOpts...)
-
-	// Build the CompleteNoActionTool.
-	completeNoActionTool := mcptools.NewCompleteNoActionTool(leaseMgr,
-		mcptools.WithCompleteNoActionLogger(logger.WithName("complete-no-action")),
-		mcptools.WithCompleteNoActionHTTPCompleter(autoMgr),
-		mcptools.WithCompleteNoActionMutexProvider(investigateTool),
-	)
-
-	// Register tools with the MCP SDK server.
-	toolDeps := mcpkg.ToolDeps{}
-	toolDeps.Investigate = mcptools.InvestigateRegistration(investigateTool, eventStore, sessionNotifier, logger)
-	toolDeps.SelectWorkflow = mcptools.SelectWorkflowRegistration(selectWfTool, logger)
-	toolDeps.CompleteNoAction = mcptools.CompleteNoActionRegistration(completeNoActionTool, logger)
-
-	mcpKeepAlive := cfg.Interactive.MCPKeepAlive
-	if mcpKeepAlive == 0 {
-		mcpKeepAlive = kaconfig.DefaultMCPKeepAlive
-	}
-	mcpSessionTimeout := cfg.Interactive.MCPSessionTimeout
-	if mcpSessionTimeout == 0 {
-		mcpSessionTimeout = kaconfig.DefaultMCPSessionTimeout
-	}
-	mcpHandler, _ := mcpkg.BootstrapMCP(mcpkg.MCPDeps{
-		AuthMiddleware: func(next http.Handler) http.Handler {
-			return kaserver.AuditAuthMiddleware(authMw.Handler(next), auditStore, logger)
-		},
-		Tools:          toolDeps,
-		EventStore:     eventStore,
-		KeepAlive:      mcpKeepAlive,
-		SessionTimeout: mcpSessionTimeout,
-	})
-
-	drainer := mcpkg.NewSessionDrainer(leaseMgr, sessionNotifier, logger.WithName("session-drainer"))
-
-	logger.Info("MCP interactive mode fully wired",
-		"investigate", true,
-		"select_workflow", true,
-		"complete_no_action", true,
-		"enrichment_in_select_workflow", enricher != nil,
-		"event_store", true,
-		"timeout_manager", true,
-		"disconnect_handler", true,
-		"reconstruction_spawner", true,
-		"notification_bus", true,
-		"session_drainer", true,
-	)
-
-	return mcpHandler, drainer
-}
-
-// noopReconstructor is a no-op ContextReconstructor used when DS is unavailable.
-type noopReconstructor struct{}
-
-func (n *noopReconstructor) Reconstruct(_ context.Context, _ string, _ string) ([]mcpkg.ConversationTurn, error) {
-	return nil, nil
 }
